Moved messages type and langs to language type file for reusability between app.tsx and test-utils
