+++ __tests__/pages/consign.test.tsx
@@ -1,6 +1,6 @@
 import ConsignPage from '@/pages/consign';
 import { render } from '@/test-utils';
-import { langs } from '@/utilities/resources';
+import { langs } from '@/types/languages';
 
 describe('ConsignPage', () => {
     it('should render page in English', () => {
