+++ __tests__/pages/invest.test.tsx
@@ -3,7 +3,7 @@ import mockRouter from 'next-router-mock';
 
 import InvestPage from '@/pages/invest';
 import { render, screen, waitFor, act } from '@/test-utils';
-import { langs } from '@/utilities/resources';
+import { langs } from '@/types/languages';
 
 jest.mock('next/router', () => jest.requireActual('next-router-mock'));
