+++ utilities/resources.ts
@@ -1,6 +1,2 @@
 export const instagramLink: string = 'https://www.instagram.com/newyorkcitymotorcycles/';
 
-export const langs = {
-    en: 'en',
-    ja: 'ja',
-};
